requirements.txt
# EmotiAI - Emotionally Intelligent Personal AI Model
# Requirements for building, training, and deploying

# Core ML/AI Libraries
transformers>=4.40.0
peft>=0.10.0
trl>=0.8.6
bitsandbytes>=0.43.1
datasets>=2.19.0
accelerate>=0.29.3

# Deep Learning Framework
torch>=2.3.0
torchvision>=0.18.0
torchaudio>=2.3.0

# Experiment Tracking
wandb>=0.17.0

# Data Processing
numpy>=1.26.4
pandas>=2.2.2
scikit-learn>=1.4.2

# Web & API
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.2
httpx>=0.27.0

# Database
sqlalchemy>=2.0.29
aiosqlite>=0.19.0

# Utilities
tqdm>=4.66.2
python-dotenv>=1.0.1
pyyaml>=6.0.1
regex>=2024.5.15
emoji>=2.11.0

# Reddit API (for data collection)
praw>=7.7.1
